refactor(core): split the core/types.ts god file into domain modules
core/types.ts mixed pure types with runtime logic for ~8 domains. Move the
runtime functions to domain-owned modules, leaving core/types.ts as types
only (1921 -> 1513 lines):

- core/cursor.ts          encodeCursor, decodeCursor
- core/nested-writes.ts   extractNestedData, isDirectNestedData
- core/computed-fields.ts applyComputedFields, applyComputedFieldsToArray
- core/aggregate.ts       parseAggregateField, parseAggregateQuery
- core/soft-delete.ts     getSoftDeleteConfig
- audit/config.ts         getAuditConfig, calculateChanges
- versioning/config.ts    getVersioningConfig
- multi-tenant/config.ts  getMultiTenantConfig, extractTenantId (switch -> map)
- endpoints/search-utils  parseSearchMode

defineModel/defineMeta and RESPONSE_ENVELOPE_CONTEXT_KEY stay in core/types.ts.
core/types.ts remains an import leaf (moved fns import only types from it), so
no import cycles. Public barrel + internal.ts re-export the same names from
their new homes — public API unchanged.

917 unit + 42 workers tests pass; build + all-package typecheck + biome + the
local-consumer example all clean.
