+++ packages/core/src/audit/index.ts
@@ -1,8 +1,8 @@
 import type { Context, Env } from 'hono';
 import { CONTEXT_KEYS } from '../core/context-keys';
-import type { AuditAction, AuditLogEntry, NormalizedAuditConfig } from '../core/types';
-import { type AuditConfig, calculateChanges, getAuditConfig } from '../core/types';
+import type { AuditAction, AuditConfig, AuditLogEntry, NormalizedAuditConfig } from '../core/types';
 import { createRegistryWithDefault } from '../storage/registry';
+import { calculateChanges, getAuditConfig } from './config';
 
 /**
  * Interface for audit log storage adapters.
