+++ packages/core/src/endpoints/create.ts
@@ -2,6 +2,7 @@ import type { Env } from 'hono';
 import { type ZodObject, type ZodRawShape, z } from 'zod';
 import { getLogger } from '../core/logger';
 import { getManagedInputExclusions, rethrowAsConstraintError } from '../core/managed-fields';
+import { extractNestedData } from '../core/nested-writes';
 import type {
   HookContext,
   HookMode,
@@ -9,7 +10,6 @@ import type {
   OpenAPIRouteSchema,
   RelationConfig,
 } from '../core/types';
-import { extractNestedData } from '../core/types';
 import { CrudEndpoint } from './base';
 import { errorResponseSchema } from './responses';
 import { type ModelObject, getSchemaFields } from './types';
