+++ tests/new-features.test.ts
@@ -13,7 +13,7 @@ import { Hono } from 'hono';
 import { fromHono, registerCrud } from 'hono-crud';
 import type { MetaInput, Model } from 'hono-crud';
 import { generateETag, matchesIfMatch, matchesIfNoneMatch } from 'hono-crud';
-import { decodeCursor, encodeCursor } from 'hono-crud/core/types';
+import { decodeCursor, encodeCursor } from 'hono-crud';
 import { CrudEventEmitter } from 'hono-crud/events/emitter';
 import { beforeEach, describe, expect, it } from 'vitest';
 import { z } from 'zod';
